moved section col2 part to their own folder
started to look messy needed restructure
