--- src/components/GalleryCarousel/carouselData.js
@@ -1,18 +1,0 @@
-export const slidesImages = [
-  {
-    src: require("../../images/galleryCarouselImgs/1.jpg"),
-    alt: "carouselExampleImg1",
-  },
-  {
-    src: require("../../images/galleryCarouselImgs/2.jpg"),
-    alt: "carouselExampleImg2",
-  },
-  {
-    src: require("../../images/galleryCarouselImgs/3.jpg"),
-    alt: "carouselExampleImg3",
-  },
-  {
-    src: require("../../images/galleryCarouselImgs/4.jpg"),
-    alt: "carouselExampleImg4",
-  },
-];
