+++ src/components/SectionsEmelents/AboutImg/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { ImgWrapper, Img } from "./AboutImgElements";
-import AboutImage from "../../images/aboutImg/aboutImg.jpg";
+import AboutImage from "../../../images/aboutImg/aboutImg.jpg";
 
 const AboutImg = () => {
   return (
