+++ src/components/SectionsEmelents/GalleryCarousel/carouselData.js
@@ -1,0 +1,18 @@
+export const slidesImages = [
+  {
+    src: require("../../../images/galleryCarouselImgs/1.jpg"),
+    alt: "carouselExampleImg1",
+  },
+  {
+    src: require("../../../images/galleryCarouselImgs/2.jpg"),
+    alt: "carouselExampleImg2",
+  },
+  {
+    src: require("../../../images/galleryCarouselImgs/3.jpg"),
+    alt: "carouselExampleImg3",
+  },
+  {
+    src: require("../../../images/galleryCarouselImgs/4.jpg"),
+    alt: "carouselExampleImg4",
+  },
+];
