+++ src/components/SectionsEmelents/GalleryCarousel/index.js
@@ -3,7 +3,7 @@ import { CarouselContainer, Img } from "./GalleryCarouselElements";
 import { slidesImages } from "./carouselData";
 import { Splide, SplideSlide } from "@splidejs/react-splide";
 import "@splidejs/splide/css";
-import styles from "../../themes/splideCssOverwrite.css";
+import styles from "../../../themes/splideCssOverwrite.css";
 
 const GalleryCarousel = () => {
   return (
