+++ src/components/SectionsEmelents/ServicesImgs/index.js
@@ -1,9 +1,9 @@
 import React from "react";
 import { ImgWrapper, Img } from "./ServicesImgElements";
-import designImg from "../../images/servicesImgs/design.jpg";
-import tatooingImg from "../../images/servicesImgs/tattoo.jpg";
-import piercingImg from "../../images/servicesImgs/piercing.jpg";
-import microdermalImg from "../../images/servicesImgs/microdermal.jpg";
+import designImg from "../../../images/servicesImgs/design.jpg";
+import tatooingImg from "../../../images/servicesImgs/tattoo.jpg";
+import piercingImg from "../../../images/servicesImgs/piercing.jpg";
+import microdermalImg from "../../../images/servicesImgs/microdermal.jpg";
 
 const ServicesImgs = () => {
   return (
