+++ src/pages/HomePage/index.js
@@ -4,11 +4,11 @@ import Sidebar from "../../components/Sidebar";
 import SettingsModal from "../../components/SettingsModal";
 import HeroVideo from "../../components/HeroSection";
 import Section from "../../components/Sections";
-import AboutImg from "../../components/AboutImg";
-import ServicesImgs from "../../components/ServicesImgs";
-import GalleryCarousel from "../../components/GalleryCarousel";
-import ContactDetails from "../../components/ContactDetails";
-import TestimonialCarousel from "../../components/TestimonialCarousel";
+import AboutImg from "../../components/SectionsEmelents/AboutImg";
+import ServicesImgs from "../../components/SectionsEmelents/ServicesImgs";
+import GalleryCarousel from "../../components/SectionsEmelents/GalleryCarousel";
+import ContactDetails from "../../components/SectionsEmelents/ContactDetails";
+import TestimonialCarousel from "../../components/SectionsEmelents/TestimonialCarousel";
 import Footer from "../../components/Footer";
 import ProgressBar from "../../components/ProgressBar";
 import { BlurAnimation } from "./HomePageElements";
